Move component to commons
