+++ src/commons/WeekDaysNames.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {StyleProp, Text, TextStyle} from 'react-native';
-import {weekDayNames} from '../../dateutils';
+import {weekDayNames} from '../dateutils';
 
 interface WeekDaysNamesProps {
   firstDay?: number;
