+++ src/expandableCalendar/WeekCalendar/new.tsx
@@ -4,7 +4,7 @@ import XDate from 'xdate';
 
 import InfiniteList from '../../infinite-list';
 import Week from '../week';
-import WeekDaysNames from '../../calendar/header/WeekDaysNames';
+import WeekDaysNames from '../../commons/WeekDaysNames';
 import {CalendarListProps} from '../../calendar-list';
 import CalendarContext from '../../expandableCalendar/Context';
 import styleConstructor from '../style';
